moving project json parser util to vscode-core
