+++ packages/salesforcedx-utils-vscode/src/cli/index.ts
@@ -33,4 +33,3 @@ export {
   DeployError
 } from './deployErrorParser';
 export { TestRunner } from './testRunner';
-export { SfdxProjectJsonParser } from './sfdxProjectJsonParser';
